MYR-832: move interface to own file
