+++ src/components-v2/atoms/customAsset/customAsset.interface.ts
@@ -1,0 +1,11 @@
+export type Props = {
+  open: boolean;
+  onClose: () => void;
+};
+
+export interface Asset {
+  rpcAddress: string;
+  unitCode: string;
+  prefixAddress: number;
+  prefixDecimal: number;
+}
